sdk/go: Support looking inside []ResourceOption
Adds a new ResourceOptions type
and a constructor for it

    func NewResourceOptions(...ResourceOption) ResourceOptions

Together, these provide a means for users
to look inside a `...ResourceOption` argument
from mocks and component resource implementations.

ResourceOptions is intended to *mostly* mirror
the internal resourceOptions struct
that the ResourceOption interface operates on.
However, it diverges from the internal representation
in a few places for a better user-facing API.
For example,

- instead of `map[string]Provider`, expose `[]Provider`
  sorted by the `string`
- instead of dependency URN information,
  expose fields analogous to the original options

Refs #11698